Adding CC bib, fixing URL, fix journals
New bibliography file for creative commons images. Not entirely sure how
I am dealing with them at this point.

Fixing journal entries that were missing or wrong. Also added the list
of journals to the makefile depedency.

Fixed the URL so that there are no longer angle brackets surrounding it.